requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
plotly==5.15.0
openai==0.27.8
anthropic==0.2.0
boto3==1.26.155
gemini-sdk==1.0.0
xai-sdk==1.0.0
writer-sdk==1.0.0